rough draft revision one
Created new Calculator GUI that utilizes input validation
Added toString() to Binary and Hexadecimal Classes
Created Validator Class
Added isInt(), isBigInt(), isDecimal(), isHex(), and isBinary() to Validator
Removed Test Class

+++ src/Controller/DecimalCalculator.java
@@ -30,7 +30,7 @@ public class DecimalCalculator extends Calculator{
     }
 
     public boolean isNum(String str){
-        try { Integer.parseInt(str); return true; }
+        try { Double.parseDouble(str); return true; }
         catch (NumberFormatException nfe){ return false; }
     }
 }
